// car-booking-fe-service/car-booking/src/models/response/booking/bookingResponse.ts
export interface CarBookingResponse {
    booking_id: number;
    car_id: number;
    car_model_brand: string;
    car_model_name: string;
    car_model_year: number;
    booking_date_time: string;
    customer_name: string;
    customer_email: string;
    customer_phone: string;
    created_by: string;
    created_at: string;
    updated_by: string;
    updated_at: string;
  }